requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
inflection==0.5.1
jsonschema==4.17.0
pyrsistent==0.19.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
